+++ xarray_ms/backend/msv2/entrypoint.py
@@ -240,7 +240,7 @@ class MSv2EntryPoint(BackendEntrypoint):
     "structure_factory",
   ]
   description = "Opens v2 CASA Measurement Sets in Xarray"
-  url = "https://link_to/your_backend/documentation"
+  url = "https://xarray-ms.readthedocs.io/"
 
   def guess_can_open(
     self, filename_or_obj: str | os.PathLike[Any] | BufferedIOBase | AbstractDataStore
@@ -412,11 +412,14 @@ class MSv2EntryPoint(BackendEntrypoint):
       ms, ninstances, auto_corrs, epoch, None, partition_columns, None
     )
 
+    # /path/to/some_name.ext -> some_name
+    ms_name, _ = os.path.splitext(os.path.basename(ms.rstrip(os.path.sep)))
+
     structure = structure_factory()
     datasets = {}
     pchunks = promote_chunks(structure, preferred_chunks)
 
-    for partition_key in structure:
+    for p, partition_key in enumerate(structure):
       ds = xarray.open_dataset(
         ms,
         drop_variables=drop_variables,
@@ -435,9 +438,9 @@ class MSv2EntryPoint(BackendEntrypoint):
 
       antenna_factory = AntennaDatasetFactory(structure_factory)
 
-      key = ",".join(f"{k}={v}" for k, v in sorted(partition_key))
-      ds.attrs["antenna_xds_link"] = f"{key}/ANTENNA"
-      datasets[key] = ds
-      datasets[f"{key}/ANTENNA"] = antenna_factory.get_dataset()
+      path = f"{ms_name}/partition_{p:03}"
+      ds.attrs["antenna_xds_link"] = f"{path}/ANTENNA"
+      datasets[path] = ds
+      datasets[f"{path}/ANTENNA"] = antenna_factory.get_dataset()
 
     return datasets